+++ src/Form/Type/FilterCalendarEventsType.php
@@ -8,19 +8,25 @@ use App\Enum\EventClassroomTypeEnum;
 use App\Form\Model\FilterCalendarEventModel;
 use App\Repository\ClassGroupRepository;
 use App\Repository\TeacherRepository;
-use Symfony\Bridge\Doctrine\Form\Type\EntityType;
 use Symfony\Component\Form\AbstractType;
 use Symfony\Component\Form\Extension\Core\Type\ChoiceType;
+use Symfony\Component\Form\Extension\Core\Type\SubmitType;
 use Symfony\Component\Form\FormBuilderInterface;
+use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\OptionsResolver\OptionsResolver;
+use Symfony\Component\Routing\RouterInterface;
 
 class FilterCalendarEventsType extends AbstractType
 {
+    public const SESSION_KEY = 'filter_calendar_events_form_data';
+
+    private RouterInterface $rs;
     private TeacherRepository $tr;
     private ClassGroupRepository $cgr;
 
-    public function __construct(TeacherRepository $tr, ClassGroupRepository $cgr)
+    public function __construct(RouterInterface $rs, TeacherRepository $tr, ClassGroupRepository $cgr)
     {
+        $this->rs = $rs;
         $this->tr = $tr;
         $this->cgr = $cgr;
     }
@@ -33,33 +39,55 @@ class FilterCalendarEventsType extends AbstractType
                 ChoiceType::class,
                 [
                     'label' => 'backend.admin.event.classroom',
+                    'placeholder' => 'backend.admin.event.classroom',
                     'choices' => EventClassroomTypeEnum::getEnumArray(),
                     'multiple' => false,
                     'expanded' => false,
-                    'required' => true,
+                    'required' => false,
                 ]
             )
             ->add(
                 'teacher',
-                EntityType::class,
+                ChoiceType::class,
                 [
                     'label' => 'backend.admin.event.teacher',
-                    'required' => true,
-                    'class' => Teacher::class,
-                    'choice_label' => 'name',
-                    'query_builder' => $this->tr->getEnabledSortedByNameQB(),
+                    'placeholder' => 'backend.admin.event.teacher',
+                    'choices' => $this->tr->getEnabledSortedByName(),
+                    'choice_value' => 'id',
+                    'choice_label' => function (?Teacher $teacher) {
+                        return $teacher ? $teacher->getName() : '';
+                    },
+                    'multiple' => false,
+                    'expanded' => false,
+                    'required' => false,
                 ]
             )
-//            ->add(
-//                'group',
-//                EntityType::class,
-//                [
-//                    'label' => 'backend.admin.event.group',
-//                    'required' => true,
-//                    'class' => ClassGroup::class,
-//                    'query_builder' => $this->cgr->getEnabledSortedByCodeQB(),
-//                ]
-//            )
+            ->add(
+                'group',
+                ChoiceType::class,
+                [
+                    'label' => 'backend.admin.event.group',
+                    'placeholder' => 'backend.admin.event.group',
+                    'choices' => $this->cgr->getEnabledSortedByCode(),
+                    'choice_value' => 'id',
+                    'choice_label' => function (?ClassGroup $classGroup) {
+                        return $classGroup ? $classGroup->__toString() : '';
+                    },
+                    'multiple' => false,
+                    'expanded' => false,
+                    'required' => false,
+                ]
+            )
+            ->add(
+                'submit',
+                SubmitType::class,
+                [
+                    'label' => 'backend.admin.filter_calendar_by',
+                    'attr' => [
+                        'class' => 'btn btn-warning pull-right',
+                    ],
+                ]
+            )
         ;
     }
 
@@ -67,6 +95,8 @@ class FilterCalendarEventsType extends AbstractType
     {
         $resolver->setDefaults(
             [
+                'method' => Request::METHOD_POST,
+                'action' => $this->rs->generate('admin_app_filter_calendar'),
                 'data_class' => FilterCalendarEventModel::class,
             ]
         );